# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tuquet Automa Core Bridge API",
    "description": "Lightweight native process launcher, WebSocket event hub, and SQLite storage bridge for Automa WebExtension.",
    "contact": {
      "name": "Tuquet Ecosystem",
      "url": "https://github.com/tuquet/tuquet-automa"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8765",
      "description": "Local Daemon Bridge Server"
    }
  ],
  "paths": {
...
    "/api/v1/jobs/{job_id}": {
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Terminate a running job",
        "description": "Cancels execution token for a job, releases the associated browser process, records terminated status, and clears state.",
        "operationId": "kill_job",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "description": "Unique job identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job terminated successfully"
          },
          "404": {
            "description": "Job not found or not active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/{job_id}/pause": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Pause a running job",
        "description": "Dispatches a pause-workflow signal to the worker, updates job status to 'paused', and broadcasts event via SSE.",
        "operationId": "pause_job",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "description": "Unique job identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job paused successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobControlResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found or not active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/resume": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Resume a paused job",
        "description": "Dispatches a resume-workflow signal to the worker, updates job status to 'running', and broadcasts event via SSE.",
        "operationId": "resume_job",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "description": "Unique job identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job resumed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobControlResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found or not active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "status"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "JobControlResponse": {
        "type": "object",
        "description": "Response after controlling job execution lifecycle (pause, resume)",
        "required": [
          "success",
          "message",
          "jobId",
          "status"
        ],
        "properties": {
          "jobId": {
            "type": "string",
            "description": "Target job ID"
          },
          "message": {
            "type": "string",
            "description": "Informational message"
          },
          "status": {
            "type": "string",
            "description": "New lifecycle status"
          },
          "success": {
            "type": "boolean",
            "description": "Operation success status"
          }
        },
        "example": {
          "jobId": "job_123",
          "message": "Job paused successfully",
          "status": "paused",
          "success": true
        }
      },
...
    }
  },
...
}